Repair Frontend-Test. Repair distorted Styling.
