+++ frontend/src/App.js
@@ -3,39 +3,19 @@ import ExplorePage from "./pages/ExplorePage";
 import styled from "styled-components/macro";
 import LoginPage from "./pages/LoginPage";
 import Header from "./components/Header";
-import { createMuiTheme, ThemeProvider } from "@material-ui/core";
-
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      light: "#cfff95",
-      main: "#9ccc65",
-      dark: "#6b9b37",
-      contrastText: "#ffffff",
-    },
-    secondary: {
-      light: "#67daff",
-      main: "#03a9f4",
-      dark: "#007ac1",
-      contrastText: "#ffffff",
-    },
-  },
-});
 
 function App() {
   return (
     <Wrapper>
-      <ThemeProvider theme={theme}>
-        <Header />
-        <Switch>
-          <Route path={"/"} exact>
-            <LoginPage />
-          </Route>
-          <Route path={"/explore"} exact>
-            <ExplorePage />
-          </Route>
-        </Switch>
-      </ThemeProvider>
+      <Header />
+      <Switch>
+        <Route path={"/"} exact>
+          <LoginPage />
+        </Route>
+        <Route path={"/explore"} exact>
+          <ExplorePage />
+        </Route>
+      </Switch>
     </Wrapper>
   );
 }
@@ -44,7 +24,9 @@ export default App;
 
 const Wrapper = styled.div`
   margin: 0;
-  height: 100vh;
+  height: 100%;
   width: 100vw;
   background-color: hotpink;
+  display: flex;
+  flex-direction: column;
 `;
